Whitespace change: see extended comment
At this point a large breaking change has been comleted on master.
Since we have many IndexedDataSource helper classes in the storage
package I have eliminated the various complexities of sublist
processing code in all the various algorithms. It is now assumed
that the api user will filter lists as needed using the helpers
and that all algorithms work on full lists.