After reading determine there was a bug in the correlation code
Gonzalez and Woods show that correlation involves the conjugation
of values. All the code I've seen on the web does not imply this.
But all those examples may have only been using real numbers. My
code can be used with complex and other numbers and we need to
maintain this use of conjugation.

So as a consequence break out a couple new classes. The correlation
and convolution codes are very similar but have important
differences.