Collapse FixedTransforms into TransformWithConstant
Also while at it make sure methods pass their values on the right.
For instance adds and mutiplies of scalars with vecs, mats, and
tensors do so on the right. This is consistent with divides and
subtracts. This has not yet been tested.

Note that all the ScaleBy algorithms might violate this and need
to be addressed yet. Also note that this might break code for
existing code that uses quaternions and octonions. I am of the
opinion this is a good decision but I guess we will see. Fixing
the ScaleBy code is going to take a lot of effort.