Make another algo change
ReplaceIf may have lost its meaning. It always does a ReplaceCopyIf.
This might cut down on the intended efficiency since it no longer
only assigns when true. When I did that, using Transforms and
multiple threads I got bad results. For an expensive to fill data
source this new change might make ReplaceIf too expensive to run.
Maybe I need a version that is not-threaded or is smarter.