Fix ThreadUtils.arrange(): much simpler and more accurate now
I swear at one time this is what the code did. And some test used
to fail. Submitting this code to the CI server and will see if some
tests starts failing again. If it does it is a problem with the
calling algorithm. I have used unit tests and hand tests to verify
that the new code is correct.